% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{adjust_concentrations}
\alias{adjust_concentrations}
\title{Correcting Responses assuming Deactivation}
\usage{
adjust_concentrations(conc, R, A_full)
}
\arguments{
\item{conc}{a vector of postive values representing concentrations (eg
micromol/L)}

\item{R}{a vector of real-valued weights used in the objective; examples
include docking scores or molecular weights}

\item{A_full}{a binary matrix.  if A_{ij}= 1, chemical i can substitute
chemical j.  The diagonal must be 1.  If A is the Id matrix, the output
adjusted concentrations will be equal to the input.}
}
\value{
a vector of adjusted concentrations, some of which may be 0
}
\description{
Given input concentrations, suppose some chemicals "deactivate" other
chemicals if they have a higher weight (eg molecular weight, docking score,
or affinity).  This function will adjust the true concentrations to reflect
how the higher weight chemicals will substitute the lower weight chemicals.
If the highest weight chemical has high enough concentration, it can dominate
and result in a mixture with only one effective chemical.  We can set rules
for which chemicals can substitute others using a matrix A_full.
}
\examples{
A_full <- matrix(1, nrow=3, ncol=3)
conc <- matrix(c(4,9, 7), nrow=3, ncol=1)
binding_affinity <- c(-2, -1, -0.5)
}
