% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RE_dose_response_MCMC.R
\name{build_replicate_matrix}
\alias{build_replicate_matrix}
\title{Builds a design matrix for replicates of dose response data for performing a
simultaneous Gibbs update for the sill, sill random effect, and intercept
random effect}
\usage{
build_replicate_matrix(v_list)
}
\arguments{
\item{v_list}{a list with a vector for each replicate. Each vector is a
sequence of coefficients with length equal to the number of doses (ie
samples), computed according to the Hill model.}
}
\value{
a matrix built from the vectors of the input to mimic a design matrix
for a linear model
}
\description{
Builds a design matrix for replicates of dose response data for performing a
simultaneous Gibbs update for the sill, sill random effect, and intercept
random effect
}
