% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_MCMC.R
\name{cluster_centers}
\alias{cluster_centers}
\title{Get the cluster results}
\usage{
cluster_centers(cluster_chain, n_top = 5, plot_hist = T)
}
\arguments{
\item{cluster_chain}{the list of MCMC samples returned by \code{\link[=DP_MCMC_fit]{DP_MCMC_fit()}}}

\item{n_top}{integer number of clusters to select, starting with the most
frequently occuring}

\item{plot_hist}{boolean to plot a histogram showing how often a cluster has
k components or centers}
}
\value{
a list with the cluster centers, standard deviations, approximated
probabilities, assignments, and the number of leftover or unused
clusterings
}
\description{
Provides the top clusters and the corresponding centers given the MCMC chains
provided by \code{\link[=DP_MCMC_fit]{DP_MCMC_fit()}}
}
\examples{
y_i <- c(-1.48, -1.4, -1.16, -1.08, -1.02, 0.14, 0.51, 0.53, 0.78)
cluster_chain <- DP_MCMC_fit(y_i)
clust_centers_w_prob <- cluster_centers(cluster_chain, n_top = 20)
}
