% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_MCMC.R
\name{convert_DP_pack_obj}
\alias{convert_DP_pack_obj}
\title{Convert output from dirichletprocess package}
\usage{
convert_DP_pack_obj(dp_fit, n_top = 5)
}
\arguments{
\item{dp_fit}{output from \code{\link[dirichletprocess:DirichletProcessGaussian]{dirichletprocess::DirichletProcessGaussian()}}}

\item{n_top}{integer number of top clusters to select}
}
\value{
a list with the same attributes as the \code{\link[=cluster_centers]{cluster_centers()}} output
}
\description{
This script converts the fitted model object from the
\code{\link[dirichletprocess:DirichletProcessGaussian]{dirichletprocess::DirichletProcessGaussian()}} to match our own
\code{\link[=cluster_centers]{cluster_centers()}} output.  This allows us to reuse our visualize functions
or use their code for the pipeline.
}
