% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_calculators.R
\name{create_mix_calc_clustered}
\alias{create_mix_calc_clustered}
\title{Mixture Response Calculator Wrapper for Manuscript, using DP}
\usage{
create_mix_calc_clustered(idx, par_list, add_RE = TRUE, unit_slopes = FALSE)
}
\arguments{
\item{idx}{Specifies which clustering to apply to the parameters}

\item{par_list}{a data frame with individual chemical dose response
parameters}

\item{add_RE}{boolean to include or exclude random effect variances}

\item{unit_slopes}{boolean to fix slopes to 1 (but still use slope
clustering) Used for special case of GCA, where slope = 1}
}
\value{
function to take a concentration vector as input and response as
output
}
\description{
A factory method that returns a function that computes the mixture response.
Samples directly from the posterior MCMC chains for the EC50 and sill
parameters. Slope is taken from the Dirichlet Process clustering.  Noise is
added to some parameters according to the random effect variance.
}
