% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{create_mix_calc_from_summary}
\alias{create_mix_calc_from_summary}
\title{These functions are wrappers for the factory method mix_function_generator.
The wrappers provide different ways of sampling the parameters.
Creates a function that computes mixture response.}
\usage{
create_mix_calc_from_summary(idx, par_list)
}
\arguments{
\item{idx}{Specifies which clustering to apply to the parameters}

\item{par_list}{Contains all estimated parameters and DP cluster options}
}
\value{
a function that takes a concentration vector as input and yields a
predicted response as output
ie a "calculator" for the mixture effect given component concentrations
}
\description{
A factory method that returns a function that computes the mixture response.
Uses summary statistics to create a random sample rather than directly
sampling from the posterior MCMC chains
}
