% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_MCMC.R
\name{DP_MCMC_fit}
\alias{DP_MCMC_fit}
\title{Dirichlet Process MCMC}
\usage{
DP_MCMC_fit(y_in, n_iter = 10000, sigma_in = 0.1)
}
\arguments{
\item{y_in}{a vector of real numbers to be clustered}

\item{n_iter}{an integer for the number of iterations, defaults to 10,000}

\item{sigma_in}{a positive integer for the variances of the clusters.  A
large value leads to fewer clusters.}
}
\value{
a list with MCMC samples for the cluster means, cluster assignments,
concentration parameter alpha, and the cluster variance sigma
}
\description{
Performs clustering on the input data according to a Dirichlet Process
Mixture model.  An implementation of Neal 1999.
}
\examples{
y_i <- c(-1.48, -1.4, -1.16, -1.08, -1.02, 0.14, 0.51, 0.53, 0.78)
DP_MCMC_fit(y_i)
}
