% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{eff_response_opt}
\alias{eff_response_opt}
\title{A factory function that returns a function with one input that can be
optimized with respect to that input. The returned function computes the
generalized concentration addition formula for the input set of chemicals and
input concentrations.}
\usage{
eff_response_opt(
  hill_inverse_list,
  conc_vec,
  synergy_const = 0,
  interval_sign = 1
)
}
\arguments{
\item{hill_inverse_list}{a list of the hill inverse functions with one
function per chemical}

\item{conc_vec}{a vector of non-negative values with the same length as the
hill_inverse_list}

\item{synergy_const}{a scaling term for synergy. Difficult to specify so it
is not used and currently set to 0, implying no synergy.}

\item{interval_sign}{1 by default and can be set to -1.  Used for
chemicals with a negative sill. Details: For numerical stability around
very small concentrations and responses, the output function exponentiates
the input response, making it impossible to invert a negative response. If
at least one chemical in the mixture has a negative sill, this parameter
should be set to -1 to check if the optimal (predicted) response is
negative.}
}
\value{
a function GCA_over_list that has a response r as a input and a
norm (measure of optimality according to GCA) as output
}
\description{
A factory function that returns a function with one input that can be
optimized with respect to that input. The returned function computes the
generalized concentration addition formula for the input set of chemicals and
input concentrations.
}
