% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox21_prep_data.R
\name{generate_mixture_subset}
\alias{generate_mixture_subset}
\title{Generate a string based on the vector of mixture indices that
satisfy the boolean input conditions}
\usage{
generate_mixture_subset(
  drop_no_effect = F,
  use_graded_filter = T,
  filter_graded_with_all_nonzero = F,
  use_binary_filter = F,
  ignore_no_effect_mix = F,
  skip_zearalenone = F
)
}
\arguments{
\item{drop_no_effect}{fix agonists for other assays to 0}

\item{use_graded_filter}{mixtures with the term "graded" in description}

\item{filter_graded_with_all_nonzero}{graded mixtures involving all chemicals}

\item{use_binary_filter}{mixtures with the term "binary" in description}

\item{ignore_no_effect_mix}{if all AR chemicals have 0 concentration}

\item{skip_zearalenone}{Data issues for ER data, skipped}
}
\description{
Generate a string based on the vector of mixture indices that
satisfy the boolean input conditions
}
