% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{get_mle_curve_fits}
\alias{get_mle_curve_fits}
\title{Maximum likelihood dose response curves}
\usage{
get_mle_curve_fits(y_i, Cx, replicate_sets)
}
\arguments{
\item{y_i}{matrix of observed dose responses (columns) for multiple chemicals
and replicates (rows)}

\item{Cx}{a matrix of the doses corresponding to the responses}

\item{replicate_sets}{a list of vectors where each vector contains the
indices of replicates for a particular chemical}
}
\value{
a matrix of parameters with a row for each chemical and columns corresponding to ()
}
\description{
Wrapper function to apply the drc (dose response curve) package to the input
data.
}
