% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{hill_function}
\alias{hill_function}
\title{Hill function}
\usage{
hill_function(a, b, c, conc)
}
\arguments{
\item{a}{the maximum effect or sill}

\item{b}{the EC50}

\item{c}{the slope}

\item{conc}{an input concentration or dose}
}
\value{
a real number prediction of the dose response
}
\description{
A simple function to compute the Hill function response given parameters and
a concentration
}
