% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{make_ggplot}
\alias{make_ggplot}
\title{Makes a ggplot of the mixture response curves}
\usage{
make_ggplot(
  Cx_axis_values,
  resp_y_values,
  curve_data,
  curve_data_GCA = NULL,
  curve_data_IA = NULL
)
}
\arguments{
\item{Cx_axis_values}{matrix of mixture concentrations}

\item{resp_y_values}{matrix of observed mixture responses}

\item{curve_data}{matrix of predicted mixture responses, with a row for each
bootstrapped calculator, RGCA method}

\item{curve_data_GCA}{matrix of predicted mixture responses, GCA method}

\item{curve_data_IA}{matrix of predicted mixture responses, IA method}
}
\value{
a ggplot
}
\description{
Creates plots showing the various methods as banded curves where the bands
are the posterior credible intervals and the observed mixture response as
points.
}
