% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{make_ggplot_many}
\alias{make_ggplot_many}
\title{Makes a ggplot of the mixture response curves}
\usage{
make_ggplot_many(Cx_axis_values, resp_y_values, curve_data_list)
}
\arguments{
\item{Cx_axis_values}{matrix of mixture concentrations}

\item{resp_y_values}{matrix of observed mixture responses}

\item{curve_data_list}{a list of matrices of predicted mixture responses.
The rows of the matrices correspond to bootstrapped calculators, and each
matrix can represent a different method.}
}
\value{
a ggplot
}
\description{
Creates plots showing the various methods as banded
curves where the bands are the posterior credible intervals and the observed
mixture response as points.  Slightly more general than the make_ggplot
function because a list of matrices can be passed rather than a separate
matrix for each prediction method.
}
