% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{plot_dummy_mixture}
\alias{plot_dummy_mixture}
\title{Wrapper function to plot dummy concentrations}
\usage{
plot_dummy_mixture(
  Cx,
  y_i,
  tot_par_list,
  replicate_sets,
  bootstrapped_calc_list,
  test_idx = 1
)
}
\arguments{
\item{Cx}{matrix of concentrations or doses}

\item{y_i}{matrix of predicted responses}

\item{tot_par_list}{list of fitted parameters}

\item{replicate_sets}{a list of vectors.  The vectors contain the indices of
replicates for a particular chemical.}

\item{bootstrapped_calc_list}{a list of lists of predictors with bootstrapped
parameters.}

\item{test_idx}{the index of the individual chemical to plot}
}
\description{
A dummy concentration refers to a mixture with only one chemical.  This
plotting function is just used to test that the random effect model produces
a good fit.
}
