% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{plot_mix_vs_individuals}
\alias{plot_mix_vs_individuals}
\title{Compare individual and Mixture response}
\usage{
plot_mix_vs_individuals(mix_idx = 34, ymax = 120)
}
\arguments{
\item{mix_idx}{integer specifying which mixture to plot}

\item{ymax}{positive real number for the plot y limit}
}
\description{
Creates a series of plots where each plot shows the mixture response with one
of the individual component responses.  The mixture response concentration is
scaled to only correspond to the concentration of the compared chemical.  For
example, if a mixture has a concentraton of 5 and one component contribute 1,
when plotted against that component, the mixture response will be shifted to
a concentration of 1.
}
\details{
This function is not optimized and has some hardcoded bounds.
}
