% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{plot_mixture_response}
\alias{plot_mixture_response}
\title{Plot the mixture response}
\usage{
plot_mixture_response(
  plot_set,
  mix_df,
  mix_conc_df,
  mix_guide,
  bootstrap_calc_list
)
}
\arguments{
\item{plot_set}{a vector of indices that specify which mixtures are predicted
and plotted.  The indices correspond to the observed mixture response data
file.}

\item{mix_df}{A data frame with the observed mixture response data}

\item{mix_conc_df}{a data frame with the concentrations of the mixture (not
split into constituents)}

\item{mix_guide}{A data frame that specifies the concentrations of the
constituent chemicals for each mixture at the highest dose, along with name
and other meta data.}

\item{bootstrap_calc_list}{a list of functions that predict a mixture
response given a concentration.  The functions are created by the factory
function mix_function_generator}
}
\description{
This method creates the manuscript plots that show the predicted median
mixture response, the credible interval bands around the median response, and
the observed mixture response with replicates.
}
