% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{plot_sample_indv}
\alias{plot_sample_indv}
\title{Plot individual dose response curves}
\usage{
plot_sample_indv(df)
}
\arguments{
\item{df}{a data frame with the prepared tox21 data from
source("Code/tox21_prep_data.R")}
}
\value{
a ggplot with four examples of the tox21 data
}
\description{
Plots a few hard-coded samples of the data for a comparison used in the
manuscript.  Shows why a random effect model is needed for the Tox21 data.
Requires the data cleaning script to be previously sourced,
source("Code/tox21_prep_data.R").
}
