% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{plot_scores}
\alias{plot_scores}
\title{Create a violin plot given a data frame with scores.}
\usage{
plot_scores(score_df, bootstrap_calc_list, method_levels = NA)
}
\arguments{
\item{score_df}{The score data frame is a matrix where the columns represent
different score calculations and prediction methods while the rows
represent different data sets.}

\item{bootstrap_calc_list}{The named list of prediction functions, only used
to extract the names.}

\item{method_levels}{Optional vector of the bootstrap_calc_list names used to
adjust the order of the methods as they appear in the plot.}
}
\description{
In this context the plot compares different mixture prediction methods using
the different scoring methods.
}
