% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{predict_mix_response}
\alias{predict_mix_response}
\title{Predict the mixture response}
\usage{
predict_mix_response(
  sampled_mix_funs,
  sampled_mix_funs_GCA,
  sampled_mix_funs_IA,
  n_dose,
  chem_conc_matr,
  default_entry = 0
)
}
\arguments{
\item{sampled_mix_funs}{list of boostrapped functions for the main method of
the manuscript, RGCA+DP, instances produced by the factory method
mix_function_generator}

\item{sampled_mix_funs_GCA}{list of boostrapped functions for the standard
method of GCA}

\item{sampled_mix_funs_IA}{list of boostrapped functions for the standard
method of IA}

\item{n_dose}{integer number of doses of the mixture}

\item{chem_conc_matr}{a matrix where the rows represent the constituent
chemicals and the columns represent the dose.  The column sum is the
mixture dose.}

\item{default_entry}{a default entry for the methods, with default=0.  Can
be set to null NA to aid in plotting.}
}
\description{
This is a convenience function that applies the lists of bootstrapped mixture
response calculators to the matrix of concentrations that define the doses
for the mixture.
}
