% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_calculators.R
\name{predict_mix_response_many}
\alias{predict_mix_response_many}
\title{Predict Responses for a List of Calculators, generic}
\usage{
predict_mix_response_many(
  n_dose,
  chem_conc_matr,
  bootstrap_calc_list,
  default_entry = 0
)
}
\arguments{
\item{n_dose}{number of mixture doses}

\item{chem_conc_matr}{a matrix where the rows represent the constituent
chemicals and the columns represent the dose.  The column sum is the
mixture dose.}

\item{bootstrap_calc_list}{a list of lists of boostrapped functions,
instances produced by the factory method mix_function_generator}

\item{default_entry}{a default entry for the methods, with default=0.  Can be
set to null NA to aid in plotting}
}
\value{
A list of matrices where each matrix has dose by column and the
bootstrapped response prediction as a row.
}
\description{
A similar function to predict_mix_response.  This is a convenience function
that takes a list of lists as input and iteratively applies the elements,
which are bootstrapped mixture repsonse calculators, to the columns of the
mixture dose matrix.  For example, three different clustering approaches can
be used for the outer list while each inner list has 50 bootstrapped mixture
response estimators.  Then the outer list has length 3 and each inner list
has length 50.
}
