% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{pull_parameters}
\alias{pull_parameters}
\title{Given an MCMC chain fit by RE_MCMC_fit, pull out the chains representing the
parameter posteriors and compute some estimates of the parameters by applying
the summary statistic function to the posterior samples.}
\usage{
pull_parameters(re_chains, summry_stat = median)
}
\arguments{
\item{re_chains}{output from RE_MCMC_fit consisting of a list of parameter
chains, where each chain is a vector of posterior samples}

\item{summry_stat}{the statistic used to summarize the posterior.  Default is
median, but can be mean or any other similar measure of the center.}
}
\value{
a list with posterior samples and summary estimates of the parameters
}
\description{
Given an MCMC chain fit by RE_MCMC_fit, pull out the chains representing the
parameter posteriors and compute some estimates of the parameters by applying
the summary statistic function to the posterior samples.
}
