% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_calculators.R
\name{pull_parameters_nimble}
\alias{pull_parameters_nimble}
\title{Organize output of the Nimble MCMC}
\usage{
pull_parameters_nimble(
  nimble_samples,
  summry_stat = stats::median,
  input_replicates = NA
)
}
\arguments{
\item{nimble_samples}{the object returned by the run_RE_nimble function found
in the inst folder.}

\item{summry_stat}{The function used to compute the summarizing statistic for
the slope parameter.  Default function is median.}

\item{input_replicates}{The optional list of indices of replicates for each
chemical.  By default, input_replicates is NA and the method assumes there
exists a global var called replicate_sets}
}
\value{
A named list of arrays where each array is a posterior thinned sample
from the nimble chain.
}
\description{
Fitting a random effect model to data using a Bayesian method like MCMC
creates posterior samples.  Our implementation with Nimble creates a large
output object that contains posterior samples after burn-in and thinning.
This function organizes the samples for convenient downstream usage.
}
