% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_calculators.R
\name{pull_summary_parameters}
\alias{pull_summary_parameters}
\title{Pull parameters from an MCMC chain}
\usage{
pull_summary_parameters(re_chains, summry_stat = stats::median)
}
\arguments{
\item{re_chains}{output from RE_MCMC_fit consisting of a list of parameter
chains, where each chain is a vector of posterior samples}

\item{summry_stat}{the statistic used to summarize the posterior.  Default is
median, but can be mean or any other similar measure of the center.}
}
\value{
a list of parameter estimates
}
\description{
Given an MCMC chain fit by RE_MCMC_fit, estimate the parameter values using a
summary statistic applied to the posterior samples.
}
