% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox21_prep_data.R
\name{read_prepared_Tox21_data}
\alias{read_prepared_Tox21_data}
\title{Read preprocessed Tox21 Input data}
\usage{
read_prepared_Tox21_data(df, mix_guide)
}
\arguments{
\item{df}{dataframe with the processed data sets from Tox21.  Includes single
chemical mixtures for computing dose responses and mixtures.}

\item{mix_guide}{dataframe with ingredients and doses for each mixture.}
}
\description{
Read preprocessed Tox21 Input data
}
