% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{visualize_clusters}
\alias{visualize_clusters}
\title{Visualize Feasible Clusters}
\usage{
visualize_clusters(slopes, clust_centers_w_prob)
}
\arguments{
\item{slopes}{vector of real slope values}

\item{clust_centers_w_prob}{list object containing top clusters, returned
from \code{\link[=cluster_centers]{cluster_centers()}}.}
}
\value{
original plot settings from command \code{\link[=par]{par()}}
}
\description{
Creates a plot that visualizes the top clusters as connected line segments.
}
