% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plots.R
\name{visualize_clusters_blocks}
\alias{visualize_clusters_blocks}
\title{Visualize Feasible Clusters}
\usage{
visualize_clusters_blocks(slopes, clust_centers_w_prob, ul = 2, ll = 0)
}
\arguments{
\item{slopes}{a vector of the slope values}

\item{clust_centers_w_prob}{a named list with the results of the dirichlet
process clustering}

\item{ul}{Upper limit for plotting the cluster values}

\item{ll}{lower limit for plotting the cluster values}
}
\value{
no return object specified other than original plot settings \code{\link[=par]{par()}}
}
\description{
Creates a plot that visualizes the top clusters as boxes surrounding the
point estimates.  Provides additional information about the ordering on the
left and the frequency of the particular clustering on the right.
}
