% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculators.R
\name{hill_invs_factry}
\alias{hill_invs_factry}
\title{Title}
\usage{
hill_invs_factry(a, b, c, max_R = 1, d = 0)
}
\arguments{
\item{a}{sill (max effect)}

\item{b}{EC50}

\item{c}{slope}

\item{max_R}{maximum effect across all chemicals}

\item{d}{minimum effect}
}
\value{
a function to compute the inverse given a mixture response R
}
\description{
Title
}
\examples{
NA
}
