% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox21_prep_data.R
\name{read_prepared_Tox21_data}
\alias{read_prepared_Tox21_data}
\title{Read preprocessed Tox21 Input data}
\usage{
read_prepared_Tox21_data(input_file = "AR-luc.txt")
}
\arguments{
\item{input_file}{processed data sets from Tox21 with mixtures, included
with package}
}
\description{
Read preprocessed Tox21 Input data
}
